refactor finalize filter and sort
